fix: Select correct lines when invoked through key
If a created a selection of lines by selecting the line in the bottom
first this tool did not worked correctly when it was invoked from using
a keymap.

This fixes that by making sure the start line is the minimum line number
and end line has the maximum line number.

Note this was not a problem when this tool was invoked as a user
command.

Thank you @haolian9 for reporting this problem!